# docs/contracts/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cerberus API",
    "description": "Cerberus encrypted vault API. Identity authentication and vault access are separate.",
    "version": "v1"
  },
  "paths": {
    "/api/accounts/me": {
      "put": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "name": "X-Cerberus-Vault-Access",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAccountOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAccountOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAccountOutputDataOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "name": "X-Cerberus-Vault-Access",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOutputDataOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAccountCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAccountCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAccountCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAccountOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAccountOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAccountOutputDataOutput"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAccountOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAccountOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAccountOutputDataOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountOutput": {
        "required": [
          "details"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/EncryptedEnvelope"
          }
        },
        "additionalProperties": false
      },
      "AccountOutputDataOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/AccountOutput"
          }
        },
        "additionalProperties": false
      },
...
      "EncryptedEnvelope": {
        "type": "object",
        "properties": {
          "format": {
            "type": "string",
            "nullable": true
          },
          "keyEpoch": {
            "type": "integer",
            "format": "int64"
          },
          "keySalt": {
            "type": "string",
            "nullable": true
          },
          "nonce": {
            "type": "string",
            "nullable": true
          },
          "ciphertext": {
            "type": "string",
            "nullable": true
          },
          "tag": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "HeimdallRegistration": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProcessOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "RegisterAccountCommand": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid"
          },
          "identity": {
            "$ref": "#/components/schemas/HeimdallRegistration"
          },
          "details": {
            "$ref": "#/components/schemas/EncryptedEnvelope"
          }
        },
        "additionalProperties": false
      },
      "RegisterAccountOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "onboardingState": {
            "type": "string",
            "nullable": true
          },
          "replayed": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RegisterAccountOutputDataOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/RegisterAccountOutput"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateAccountCommand": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          },
          "details": {
            "$ref": "#/components/schemas/EncryptedEnvelope"
          }
        },
        "additionalProperties": false
      },
      "UpdateAccountOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "UpdateAccountOutputDataOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/UpdateAccountOutput"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}